Revert "Revert "Merge remote-tracking branch 'origin/master' into Menu""
This reverts commit ebc183d119fff74131f7a047e3dc8a5913110b8e.

+++ Assets/Scripts/Controllers/HealthController.cs
@@ -9,15 +9,15 @@ namespace Controllers
     [RequireComponent(typeof(Actor))]
     public class HealthController : MonoBehaviour, IDamageable
     {
-        public int MaxHealth => GetComponent<Actor>().ActorStats.MaxHealth;
-        [SerializeField] private int _currentHealth;
+        public float MaxHealth => GetComponent<Actor>().ActorStats.MaxHealth;
+        [SerializeField] private float _currentHealth;
 
         private void Start()
         {
             _currentHealth = MaxHealth;
         }
 
-        public void TakeDamage(int damage)
+        public void TakeDamage(float damage)
         {
             _currentHealth -= damage;
