+++ Assets/Scripts/Strategy/IHittable.cs
@@ -2,6 +2,6 @@ namespace Strategy
 {
     public interface IHittable
     {
-        void Hit(int damage);
+        void Hit(float damage);
     }
 }
